Incoming file via relative path
Assuming a user downloads the BMt-main directory directly from github and sets it as a working directory in vscode, they just need to put the file in the same place as the sample data file and change the file name in the existing code.

Relative paths are used and the code for the import part of the file is wrapped in function form.